only auto-uninstall addons of the ploneintranet-namespace

+++ src/ploneintranet/suite/uninstall.py
@@ -5,8 +5,10 @@ import logging
 
 log = logging.getLogger(__name__)
 
-
-# dependencies of our dependencies that will be removed
+# These add-ons will be uninstalled. They are either dependencies
+# of ploneintranet.suite that are not from the ploneintranet-namespace
+# or are dependencies of dependencies of ploneintranet.suite that make little
+# or no sense out side its context
 ADDITIONAL_DEPENDENCIES = [
     'slc.docconv',
     'collective.workspace',
@@ -22,7 +24,8 @@ def uninstall(context):
 
 
 def uninstall_dependencies(portal):
-    # Uninstall all dependencies of ploneintranet and its dependencies
+    # Uninstall dependencies of ploneintranet from the
+    # ploneinstranet-namespace plus some additional add-ons.
     # We do not check if the product is installed since it may be
     # INonInstallable
     qi = api.portal.get_tool('portal_quickinstaller')
@@ -32,8 +35,9 @@ def uninstall_dependencies(portal):
     dependencies.extend(ADDITIONAL_DEPENDENCIES)
     for dependency in dependencies:
         dependency = str(dependency.split('profile-')[-1].split(':')[0])
-        try:
-            qi.uninstallProducts([dependency])
-            log.info('Uninstalled dependency "%s"' % dependency)
-        except (AttributeError, BadRequest):
-            pass
+        if dependency.startswith('ploneintranet.'):
+            try:
+                qi.uninstallProducts([dependency])
+                log.info('Uninstalled dependency "%s"' % dependency)
+            except (AttributeError, BadRequest):
+                pass
